refactor(pyats): improve test structure and rename TestExecutionPlan
Refactor discovery tests for better maintainability:
- Move test_test_type_resolver.py to tests/unit/pyats_core/discovery/
- Extract helpers.py with shared FIXTURES_DIR and create_mock_path
- Split groups extraction tests into separate test_groups_extraction.py
- Parametrize tests to reduce duplication (475→301 lines in main test file)
- Use _UNUSED_TEST_ROOT for tests using mock paths vs FIXTURES_DIR for real fixtures

Rename TestExecutionPlan to PyatsDiscoveryResult for clarity:
- Better reflects its purpose as discovery output, not execution context
- Update all references in discovery module and tests
